Add pluginfactory to create easily plugin from golang
This is a fairly simple factory implementation that allows to create
generic plugins reacting to events.

The PluginFactory's Run typically would be called with os.Args (first
argument is the event type, and the second is the payload).

Finally, the writer can be used to write to e.g. Stdout